+++ test/test_time.py
@@ -1,4 +1,4 @@
-from torch4keras.snippets import timeit, Timeit
+from torch4keras.snippets import timeit, Timeit, Timeit2
 import time
 
 
@@ -13,13 +13,13 @@ func()
 with Timeit() as ti:
     for i in range(10):
         time.sleep(0.1)
-        ti.lap(prefix=i, restart=False)  # 统计累计耗时
+        ti.lap(name=i, reset=False)  # 统计累计耗时
 
 # 上下文管理器 - 统计每段速度
 with Timeit() as ti:
     for i in range(10):
         time.sleep(0.1)
-        ti.lap(count=10, prefix=i, restart=True)
+        ti.lap(count=10, name=i, reset=True)
     ti(10) # 统计速度
 
 
@@ -27,5 +27,16 @@ with Timeit() as ti:
 with Timeit() as ti:
     for i in range(10):
         time.sleep(0.1)
-        ti.lap(prefix=i, restart=True)
+        ti.lap(name=i, reset=True)
     ti(10) # 统计速度
+
+ti = Timeit2()
+for i in range(10):
+    time.sleep(0.1)
+    ti.lap(name=i)
+
+for i in range(10):
+    time.sleep(0.1)
+    ti.lap(name=i)
+ti.end() # 打印时长
+